Stats: Remove old statsModule and replace it with connected module
